Rename view controllers to be more understandable
- MainViewController -> InitialViewController
- PairViewController -> MainViewController
- SettingViewController -> SettingsViewController
