+++ Smartpad-macOS/Smartpad-macOS/Controller/MainViewController.swift
@@ -1,74 +1,56 @@
 //
-//  ViewController.swift
-//  desktop-client-mac
+//  MainViewController.swift
+//  Smartpad-macOS
 //
-//  Created by alireza azimi on 2022-01-14.
+//  Created by Alireza Azimi on 2022-03-19.
 //
 
+import Foundation
 import Cocoa
-import MultipeerConnectivity
 import CoreGraphics
 
-
-
-class ViewController: NSViewController{
-   
-    @IBOutlet weak var deviceName: NSTextField!
+class MainViewController: NSViewController {    
+    @IBOutlet weak var pairingLabel: NSTextField!
+    var connectionManager: ConnectionManager!
+    var deviceList: [String] = []
+    var selectedDevice = ""
     var connData: ConnectionData!
-    var peerID: MCPeerID!
-    var p2pSession: MCSession?
-    var mcAdvertiserAssistant: MCAdvertiserAssistant?
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        connData = ConnectionData()        
-    }
-    
-    override func viewDidAppear() {
-        self.view.window?.title = "Smartpad"
+        connectionManager = ConnectionManager()
+        connData = ConnectionData()
+        connectionManager.pairVC = self
 
-        /* Don't allow resizing the window */
-        super.view.window?.styleMask.remove(.resizable)
-
-        /* The settings menu itme should be disabled until the initial setup is complete */
+        /* Now that we are in the main view controller, the settings button should be enabled */
         if #available(macOS 12.0, *) {
-            (NSApp.delegate as! AppDelegate).setSettingsEnabled(isEnabled: false)
-        }
-
-        /* If an identifier was already set, transition to the main view */
-        if (connData.getDeviceName() != "") {
-            print("Device name: ", connData.getDeviceName())
-            if let controller = self.storyboard?.instantiateController(withIdentifier: "PairView") as? PairViewController {
-                self.view.window?.contentViewController = controller
-            }
+            (NSApp.delegate as! AppDelegate).setSettingsEnabled(isEnabled: true)
         }
     }
-    
-    @IBAction func submitDeviceName(_ sender: NSButton) {
-        let val = deviceName.stringValue
-        print(val)
-        if val != ""{
-            self.connData.setDeviceName(name: val)
-            if let controller = self.storyboard?.instantiateController(withIdentifier: "PairView") as? PairViewController {
-        self.view.window?.contentViewController = controller
-        }
-        } else {
-            let alert = NSAlert()
-
-            alert.messageText = "Name cannot be empty"
-
-            alert.addButton(withTitle: "OK")
-            alert.alertStyle = .warning
-            alert.runModal()
-        }
-        
+    func setPairLabel(label: String){
+        pairingLabel.stringValue = label
     }
     
+    func updateConnectionView(status: ConnStatus){
+        (self.view as! MainView).status = status
+        self.view.setNeedsDisplay(NSRect(x: 0,y: 0,width: 500,height: 500))
+    }
+    
+    @IBAction func pairButtonSelected(_ sender: NSButton) {
+        /* The "pair" button was pressed */
+        connectionManager.startJoining()
+    }
 
-    override var representedObject: Any? {
-        didSet {
-        // Update the view, if already loaded.
-        }
+    @IBAction func settingsButtonSelected(_ sender: NSButton) {
+        /* The "settings" button was pressed */
+        showSettingsPage()
+    }
+
+    func showSettingsPage() {
+        /* This is triggered either by the settings button or through the settings menu option in the Mac menu bar */
+        let settingsVc = storyboard?.instantiateController(withIdentifier: "settings") as! SettingsViewController
+
+        self.presentAsSheet(settingsVc)
     }
 }
